Support concrete types and custom types
Previously, you could only scan into a `*geom.T`. Now, you can scan into
a `*geom.T`, or a `*geom.Polygon` (with runtime checking), or any type
that implements the `GeomScanner` interface. This enables end users to
define their own wrapper types for interoperability with other systems,
such as ORMs that require specific methods to be defined on the types
that are used on model fields.

The public interface follows the [extensibility pattern described in the
documentation for `pgtype`][1], which is also used for first-party
`pgtype` codecs: e.g., `TimeCodec` uses `TimeScanner` and `TimeValuer`.

[1]: https://pkg.go.dev/github.com/jackc/pgx/v5/pgtype#hdr-Extending_Existing_PostgreSQL_Type_Support

Test Plan:
Unit tests included. I couldn't figure out how to get encoding tests to
run in both text and binary formats, but I did manually verify that if
you change the preferred format from binary to text then the appropriate
codepath is exercised and the test still passes.